HAL1 Hack disable UID active check
- This is a temporary hack - I expect LOS will make a proper fix eventually.

The check to isUidActive (introduced in P) is expecting to be called from the cameraserver but with the Camera HAL1
patches, it is being called from the mediaserver. This triggers a permissions denial and stops the camera from working.

So, disable this check for HAL1 devices to allow the camera to function.

Change-Id: Ibe5d5e14cc7efde54c20ab961d6d452ad5dc367f